% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angularTransform.R
\name{angularTransform}
\alias{angularTransform}
\title{Angular transformation for percentages}
\usage{
angularTransform(percentage)
}
\arguments{
\item{percentage}{A number.}
}
\value{
A number.
}
\description{
Angular transformation for percentages
}
\examples{
angularTransform(25)
angularTransform(35.2)
}
