% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L.minor}
\alias{L.minor}
\title{Enzyme kinetics}
\format{A data frame with 8 observations on the following 2 variables.
\describe{
 \item{conc}{a numeric vector} 
 \item{rate}{a numeric vector} 
}}
\usage{data("L.minor")}
\source{
Cedergreen, N. and Madsen, T. V. (2002) Nitrogen uptake by the
floating macrophyte \emph{Lemna minor}, \emph{New Phytologist}, \bold{155},
285--292.
}
\description{
Enzyme kinetics
}
\keyword{datasets}
